Blackjack
Second milestone project of the course - create a game of blackjack. This was quite a long piece of work, taking me about a week to complete. I enjoyed translating the rules of blackjack in to the logical rules you can build in Python. There were a lot of steps to complete for this so to help me keep track I created and worked through the list below.

#Variable Players: Y
#Player Names Retreived: Y
#Initial Bets Taken: Y
#Initial Card Deal: Y
#Conversion of Cards Dealt to Numerical Value: Y
#Natural Check: Y
#Stand/Hit/Bust: Y
#Calculating Bet Result: Y
#Confirming Whether Player Out of Game: Y
#Option to Leave Table: Y
#Option to Join Table: Y
#New Round: Y
#New Game: Y